+++ src/cli/eject.ts
@@ -41,7 +41,8 @@ export const makeIntegration = async (handoff: Handoff) => {
 
   // perform integration ejection
   const integrationPath = getPathToIntegration(handoff, true);
-  fs.copySync(integrationPath, workingPath, { overwrite: false });
+  fs.copySync(integrationPath, workingPath, { overwrite: handoff.force ? true : false });
+  if (handoff.force) handoff.force = false;
   await buildIntegrationOnly(handoff);
   console.log(chalk.green(`Integration has been successfully created! Path: ${workingPath}`));
   return handoff;